Move templateStore and taskPopupStore
