+++ src/lib/Reusable/IconTaskElement.svelte
@@ -98,8 +98,7 @@
 <script>
   // Assumes `task` is hydrated
   import { getTrueY } from '/src/utils/core.js'
-  import { grabOffset, activeDragItem } from '/src/store'
-  import { openTaskPopup } from '/src/store/taskPopupStore.js'
+  import { grabOffset, activeDragItem, openTaskPopup } from '/src/store'
   import DoodleIcon from '$lib/Reusable/DoodleIcon.svelte'
   import { pixelsPerHour } from '/src/lib/Calendar/store.js'
   import { updateTaskNode } from '/src/db/crud.js'
