+++ src/lib/Reusable/PhotoTaskElement.svelte
@@ -96,8 +96,7 @@
 <script>
  // Assumes `task` is hydrated
  import { getTrueY } from '/src/utils/core.js'
- import { grabOffset, activeDragItem } from '/src/store'
- import { openTaskPopup } from '/src/store/taskPopupStore.js'
+ import { grabOffset, activeDragItem, openTaskPopup } from '/src/store'
  import { lazyCallable } from '/src/utils/actions.js'
  import { pixelsPerHour } from '/src/lib/Calendar/store.js'
  import { updateTaskNode } from '/src/db/crud.js'
