+++ src/lib/Reusable/TaskElement.svelte
@@ -110,8 +110,7 @@
 <script>
   // Assumes `task` is hydrated
   import { getTrueY } from '/src/utils/core.js'
-  import { grabOffset, activeDragItem } from '/src/store'
-  import { openTaskPopup } from '/src/store/taskPopupStore.js'
+  import { grabOffset, activeDragItem, openTaskPopup } from '/src/store'
   import Checkbox from './Checkbox.svelte'
   import { pixelsPerHour, treesByID } from '/src/lib/Calendar/store.js'
   import { updateTaskNode } from '/src/db/crud.js'
