+++ src/lib/TaskPopup/TaskPopup.svelte
@@ -5,10 +5,11 @@
     mostRecentlyCompletedTaskID, 
     defaultPhotoLayout, 
     photoLayoutOptions, 
-    getIconForLayout, 
+    getIconForLayout,
+    clickedTaskID,
+    closeTaskPopup
   } from '/src/store'
   import { treesByID } from '/src/lib/Calendar/store.js'
-  import { clickedTaskID, closeTaskPopup } from '/src/store/taskPopupStore.js'
   import { createDebouncedFunction } from '/src/utils/core.js'
   import RecursiveBulletPoint from '$lib/TaskPopup/RecursiveBulletPoint.svelte'
   import UXFormTextArea from '$lib/TaskPopup/UXFormTextArea.svelte'
