+++ src/lib/Templates/Templates.svelte
@@ -1,7 +1,8 @@
 <script>
   import TemplateColumn from './TemplateColumn.svelte'
   import { onMount } from 'svelte'
-  import { user, templates, calendarTasks } from '/src/store'
+  import { user, calendarTasks } from '/src/store'
+  import { templates } from './store.js'
   import Templates from '/src/db/models/Templates'
   import { filterByType } from './utils.js'
