+++ src/lib/Templates/components/EditTemplatePopup/EditTemplatePopup.svelte
@@ -2,7 +2,8 @@
   import PeriodicInput from '$lib/Templates/components/EditTemplatePopup/PeriodicInput.svelte'
   import YearlyInput from '$lib/Templates/components/EditTemplatePopup/YearlyInput.svelte'
   import EditTime from '$lib/Templates/components/EditTemplatePopup/EditTime.svelte'
-  import { user, updateTemplate, deleteTemplate, doodleIcons } from '/src/store'
+  import { user, doodleIcons } from '/src/store'
+  import { updateTemplate, deleteTemplate } from '/src/lib/Templates/store.js'
   import Templates from '/src/db/models/Templates/index.js'
   import { onMount } from 'svelte'
   import { createDebouncedFunction } from '/src/utils/core.js'
