+++ src/lib/Templates/store.js
@@ -1,8 +1,8 @@
 import { writable, get } from 'svelte/store'
-import Templates from '../db/models/Templates/index.js'
-import { Template } from '../db/schemas';
-import { user } from './index.js'
-import './themes'
+import Templates from '/src/db/models/Templates/index.js'
+import { Template } from '/src/db/schemas'
+import { user } from '/src/store/userStore.js'
+import '/src/store/themes'
 
 export const templates = writable([])
 
@@ -13,14 +13,14 @@ export function deleteTemplate({ templateID }) {
 }
 
 export async function updateTemplate({ templateID, keyValueChanges, oldTemplate }) {
-  const currentUser = get(user);
+  const currentUser = get(user)
   const newTemplate = buildNewTemplate({ oldTemplate, keyValueChanges })
-  Template.parse(newTemplate);
+  Template.parse(newTemplate)
   templates.update((templates) => templates.map((template) =>
     template.id === templateID ? newTemplate : template
   ))
   if (oldTemplate.crontab === '') {
-    return updateQuickTasks({ templateID, newTemplate, keyValueChanges, userID: currentUser.uid });
+    return updateQuickTasks({ templateID, newTemplate, keyValueChanges, userID: currentUser.uid })
   }
   const hydratedTasks = await Templates.updateWithTasks({
     userID: currentUser.uid,
@@ -40,6 +40,6 @@ function buildNewTemplate({ oldTemplate, keyValueChanges }) {
   delete newTemplate.userID
   delete newTemplate.totalMinutesSpent
   delete newTemplate.totalTasksCompleted
-  Template.parse(newTemplate);
+  Template.parse(newTemplate)
   return newTemplate
-}
+}
